requirements.txt
# Core dependencies for web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Reddit API
praw>=7.7.0

# Twitter API (using tweepy)
tweepy>=4.14.0

# Sentiment Analysis
textblob>=0.17.1
vaderSentiment>=3.3.2
nltk>=3.8.1
transformers>=4.35.0
torch>=2.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Stock price data
yfinance>=0.2.18

# Database
# sqlite3 is built into Python
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Scheduling and automation
schedule>=1.2.0
celery>=5.3.0

# Visualization (for future reporting)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0

# Logging
loguru>=0.7.0

# HTTP client improvements
httpx>=0.25.0
